+++ cargo-rbmt/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 
 * Fix discrepancy between `nightly` toolchain commit date vs. publish date.
+* Add `--baseline` functionality to the `run` command to run per-commit.
+* Push `--lockfile` arg to be per-subcommand since about half of the subcommands don't use it.
 * Add version check at runtime if `workspace.metadata.rbmt.version` is set to a semver or hash value to ensure the right version of `cargo-rbmt` is running.
 
 # [0.4.1] - 2026-06-24